dnase filter added

+++ data/peaks/scripts/download_and_process_one.sh
@@ -5,6 +5,11 @@ if [ ! -f $1-H3K27ac.narrowPeak.gz ]; then
     echo "$1-H3K27ac.narrowPeak.gz not found!"
     exit 1
 fi
+wget -O $1-DNase.macs2.narrowPeak.gz "https://egg2.wustl.edu/roadmap/data/byFileType/peaks/consolidated/narrowPeak/$1-DNase.macs2.narrowPeak.gz"
+if [ ! -f $1-DNase.macs2.narrowPeak.gz ]; then
+    echo "$1-DNase.macs2.narrowPeak.gz not found!"
+    exit 1
+fi
 wget -O $1-H3K4me3.narrowPeak.gz "https://egg2.wustl.edu/roadmap/data/byFileType/peaks/consolidated/narrowPeak/$1-H3K4me3.narrowPeak.gz"
 if [ ! -f $1-H3K4me3.narrowPeak.gz ]; then
     echo "$1-H3K4me3.narrowPeak.gz not found!"
@@ -28,12 +33,12 @@ echo "done extracting"
 
 find . -empty -delete
 
-if [ ! -f $1-H3K27me3.narrowPeak ]; then
-    echo "$1-H3K27me3.narrowPeak not found!"
-    exit 1
-fi
 if [ ! -f $1-H3K27ac.narrowPeak ]; then
     echo "$1-H3K27ac.narrowPeak not found!"
+    exit 1
+fi
+if [ ! -f $1-DNase.macs2.narrowPeak ]; then
+    echo "$1-H3K27me3.narrowPeak not found!"
     exit 1
 fi
 if [ ! -f $1-H3K4me1.narrowPeak ]; then
@@ -44,15 +49,19 @@ if [ ! -f $1-H3K4me3.narrowPeak ]; then
     echo "$1-H3K4me3.narrowPeak not found!"
     exit 1
 fi
+if [ ! -f $1-H3K27me3.narrowPeak ]; then
+    echo "$1-H3K27me3.narrowPeak not found!"
+    exit 1
+fi
 echo "asserted files existing"
 
 # refine with background
 
-bedtools sort -faidx hg19.chrom.sizes -i $1-H3K27me3.narrowPeak > tmp.1.narrowPeak
+bedtools sort -faidx hg19.chrom.sizes -i $1-H3K27ac.narrowPeak > tmp.1.narrowPeak
 bedtools complement  -i tmp.1.narrowPeak -g hg19.chrom.sizes > tmp.2.narrowPeak
 bedtools intersect -a tmp.2.narrowPeak -b background.narrowPeak > tmp.3.narrowPeak
 mv -f tmp.3.narrowPeak background.narrowPeak
-bedtools sort -faidx hg19.chrom.sizes -i $1-H3K27ac.narrowPeak > tmp.1.narrowPeak
+bedtools sort -faidx hg19.chrom.sizes -i $1-DNase.macs2.narrowPeak > tmp.1.narrowPeak
 bedtools complement  -i tmp.1.narrowPeak -g hg19.chrom.sizes > tmp.2.narrowPeak
 bedtools intersect -a tmp.2.narrowPeak -b background.narrowPeak > tmp.3.narrowPeak
 mv -f tmp.3.narrowPeak background.narrowPeak
@@ -61,6 +70,10 @@ bedtools complement  -i tmp.1.narrowPeak -g hg19.chrom.sizes > tmp.2.narrowPeak
 bedtools intersect -a tmp.2.narrowPeak -b background.narrowPeak > tmp.3.narrowPeak
 mv -f tmp.3.narrowPeak background.narrowPeak
 bedtools sort -faidx hg19.chrom.sizes -i $1-H3K4me3.narrowPeak > tmp.1.narrowPeak
+bedtools complement  -i tmp.1.narrowPeak -g hg19.chrom.sizes > tmp.2.narrowPeak
+bedtools intersect -a tmp.2.narrowPeak -b background.narrowPeak > tmp.3.narrowPeak
+mv -f tmp.3.narrowPeak background.narrowPeak
+bedtools sort -faidx hg19.chrom.sizes -i $1-H3K27me3.narrowPeak > tmp.1.narrowPeak
 bedtools complement  -i tmp.1.narrowPeak -g hg19.chrom.sizes > tmp.2.narrowPeak
 bedtools intersect -a tmp.2.narrowPeak -b background.narrowPeak > tmp.3.narrowPeak
 mv -f tmp.3.narrowPeak background.narrowPeak
@@ -78,6 +91,9 @@ echo "bedtools run on background completed"
 
 #take only H3k27ac peaks with H3k4me1 peaks
 bedtools intersect -a $1-H3K27ac.narrowPeak -b $1-H3K4me1.narrowPeak -wa >$1-H3K27ac.cleaned.narrowPeak
+
+#take only peaks with DNASE-I peaks
+bedtools intersect -a $1-H3K27ac.narrowPeak -b $1-DNase.macs2.narrowPeak -wa >$1-H3K27ac.cleaned.narrowPeak
 
 # sort -V -k 1,3 "total_merged.narrowPeak" | sortBed | tee total_merged.tmp.narrowPeak | sort -c -k1,1 -k2,2n || true
 # # bedtools sort -i total_merged.narrowPeak > total_merged.tmp.narrowPeak
